Move serviceWorker config into own directory so that only index.js remains in root.
